src/backend/requirements.txt

fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
azure-openai>=1.0.0
pydantic>=2.0.0
